% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_plot.R
\name{write_plot}
\alias{write_plot}
\title{Write Plot}
\usage{
write_plot(plot_handle, width = 800, height = 600, wd)
}
\arguments{
\item{plot_handle}{plot handle. Required}

\item{width}{width of png plot output, in px. Default to 800}

\item{height}{width of png plot output, in px. Default to 600}

\item{wd}{the path to the output directory. Required.}
}
\description{
This function allows you to save a ggplot plot handle as png.
Prints the location of the png plot.
}
\examples{
write_plot(my_cool_plot, /usr/project/results)
write_plot(my_cool_plot, /usr/project/results,width = 1600,height = 400)
}
\keyword{plot}
